@@ -1,7 +1,16 @@
 ﻿namespace Domain.Common;
+
 public abstract class DomainEntity
 {
+    public DomainEntity() { }
+    protected DomainEntity(DateTime? currentDate)
+    {
+        CreatedAt = currentDate;
+        UpdatedAt = currentDate;
+    }
+
     public int Id { get; set; } = 0;
-    public DateTime? CreatedAt { get; set; }
-    public DateTime? UpdatedAt { get; set; }
+    public DateTime? CreatedAt { get; private set; }
+    public DateTime? UpdatedAt { get; private set; }
+    public void SetUpdatedAt() => UpdatedAt = DateTime.UtcNow;
 }
